+++ R/BinaryPowerBSSR.R
@@ -6,8 +6,6 @@
 #' one-sided or a two-sided alternative, under either a restricted or an unrestricted
 #' design rule.
 #'
-#' @param asmd.p1 Assumed proportion of responders for group 1
-#' @param asmd.p2 Assumed proportion of responders for group 2
 #' @param p Vector of true pooled proportions of responders from both groups
 #' @param Delta.A Assumed treatment effect (risk difference)
 #' @param Delta.T True treatment effect (risk difference)
@@ -61,7 +59,7 @@
 #' @examples
 #' # Small BSSR calculation with the chi-squared test
 #' BinaryPowerBSSR(
-#'   asmd.p1 = 0.6, asmd.p2 = 0.3, p = 0.45,
+#'   p = 0.45,
 #'   Delta.A = 0.3, Delta.T = 0.3,
 #'   N1 = 5, N2 = 5, omega = 0.5, r = 1,
 #'   alpha = 0.025, tar.power = 0.8, Test = 'Chisq'
@@ -69,7 +67,7 @@
 #'
 #' \donttest{
 #' res <- BinaryPowerBSSR(
-#'   asmd.p1 = 0.45, asmd.p2 = 0.09, p = seq(0.14, 0.23, by = 0.01),
+#'   p = seq(0.19, 0.37, by = 0.03),
 #'   Delta.A = 0.36, Delta.T = 0.36,
 #'   N1 = 24, N2 = 24, omega = 0.5, r = 1,
 #'   alpha = 0.025, tar.power = 0.8, Test = 'Z-pool'
@@ -82,7 +80,7 @@
 #' @export
 #' @import fpCompare
 #' @importFrom stats dbinom
-BinaryPowerBSSR <- function(asmd.p1, asmd.p2, p, Delta.A, Delta.T, N1, N2, omega, r,
+BinaryPowerBSSR <- function(p, Delta.A, Delta.T, N1, N2, omega, r,
                             alpha, tar.power, Test, restricted = FALSE,
                             alternative = c('greater', 'two.sided'),
                             tsmethod = c('minlike', 'central'),