+++ R/print.bbssr_powerbssr.R
@@ -9,7 +9,7 @@
 #'
 #' @examples
 #' res <- BinaryPowerBSSR(
-#'   asmd.p1 = 0.6, asmd.p2 = 0.3, p = 0.45,
+#'   p = 0.45,
 #'   Delta.A = 0.3, Delta.T = 0.3,
 #'   N1 = 5, N2 = 5, omega = 0.5, r = 1,
 #'   alpha = 0.025, tar.power = 0.8, Test = 'Chisq'