+++ doc/bbssr-introduction.R
@@ -54,7 +54,6 @@ BinarySampleSize(p1 = 0.45, p2 = 0.09, r = 1, alpha = 0.025,
 
 ## ----bssr-power---------------------------------------------------------------
 res <- BinaryPowerBSSR(
-  asmd.p1 = 0.45, asmd.p2 = 0.09,
   p = seq(0.19, 0.37, by = 0.03),
   Delta.A = 0.36, Delta.T = 0.36,
   N1 = 24, N2 = 24, omega = 0.5, r = 1,
@@ -67,7 +66,7 @@ plot(res)
 
 ## ----rules--------------------------------------------------------------------
 args <- list(
-  asmd.p1 = 0.45, asmd.p2 = 0.09, p = seq(0.19, 0.37, by = 0.06),
+  p = seq(0.19, 0.37, by = 0.06),
   Delta.A = 0.36, Delta.T = 0.36, N1 = 24, N2 = 24, omega = 0.5, r = 1,
   alpha = 0.025, tar.power = 0.8, Test = 'Chisq'
 )